Change role separation to more 'like DDD'. Customer domain model have its own methods

+++ app/modules/customers/application/services/customer_service.py
@@ -1,7 +1,6 @@
 from app.modules.customers.domain.enums.customer_status_enum import CustomerStatusEnum
 from app.modules.customers.domain.exceptions.customer_exceptions import (
     CustomerAlreadyBlockedException,
-    CustomerAlreadyUnlockedException,
     CustomerNotFoundException,
     DriverLicenseInDatabaseException,
     PhoneNumberInDatabaseException,
@@ -19,11 +18,6 @@ class CustomerService(ICustomerService):
 
     def __init__(self, customer_repository: ICustomerRepository) -> None:
         self._customer_repository = customer_repository
-
-    # Helper
-    def _check_customer_blocked(self, customer: Customer, customer_id: int) -> None:
-        if customer.status == CustomerStatusEnum.BLOCKED:
-            raise CustomerAlreadyBlockedException(customer_id=customer_id)
 
     # Helper gets customer and checks existence
     async def _get_customer_and_check(self, customer_id: int) -> Customer:
@@ -46,8 +40,8 @@ class CustomerService(ICustomerService):
     # Adds new customer
     async def add_customer(self, create_customer_dto: CreateCustomerDTO) -> int:
         new_id: int = await self._customer_repository.add_customer(
-            customer=Customer(
-                id=-1,
+            customer=Customer.create(
+                id=None,
                 name=create_customer_dto.name,
                 last_name=create_customer_dto.last_name,
                 phone_number=create_customer_dto.phone_number,
@@ -61,7 +55,8 @@ class CustomerService(ICustomerService):
     # Deletes customer
     async def delete_customer(self, customer_id: int) -> int:
         customer: Customer = await self._get_customer_and_check(customer_id=customer_id)
-        self._check_customer_blocked(customer=customer, customer_id=customer_id)
+        if customer.status == CustomerStatusEnum.BLOCKED:
+            raise CustomerAlreadyBlockedException(customer_id=customer_id)
 
         await self._customer_repository.delete_customer(customer_id=customer_id)
         return customer_id
@@ -72,10 +67,12 @@ class CustomerService(ICustomerService):
     ) -> int:
         customer: Customer = await self._get_customer_and_check(customer_id=customer_id)
 
-        customer.name = update_customer_dto.name
-        customer.last_name = update_customer_dto.last_name
-        customer.phone_number = update_customer_dto.phone_number
-        customer.driver_license_id = update_customer_dto.driver_license_id
+        customer.update(
+            name=update_customer_dto.name,
+            last_name=update_customer_dto.last_name,
+            phone_number=update_customer_dto.phone_number,
+            driver_license_id=update_customer_dto.driver_license_id,
+        )
 
         try:
             await self._customer_repository.update_customer(
@@ -92,21 +89,17 @@ class CustomerService(ICustomerService):
     # Blocks customer
     async def block_customer(self, customer_id: int) -> int:
         customer: Customer = await self._get_customer_and_check(customer_id=customer_id)
-        self._check_customer_blocked(customer=customer, customer_id=customer_id)
+        customer.block()
 
-        await self._customer_repository.change_status_customer(
-            customer_id=customer_id, new_status=CustomerStatusEnum.BLOCKED.value
+        return await self._customer_repository.change_status_customer(
+            customer_id=customer_id, customer=customer
         )
-        return customer_id
 
     # Unlocks customer
     async def unlock_customer(self, customer_id: int) -> int:
         customer: Customer = await self._get_customer_and_check(customer_id=customer_id)
+        customer.unlock()
 
-        if customer.status == CustomerStatusEnum.UNLOCKED:
-            raise CustomerAlreadyUnlockedException(customer_id=customer_id)
-
-        await self._customer_repository.change_status_customer(
-            customer_id=customer_id, new_status=CustomerStatusEnum.UNLOCKED.value
+        return await self._customer_repository.change_status_customer(
+            customer_id=customer_id, customer=customer
         )
-        return customer_id
